requirements.txt
black==24.10.0
click==8.1.7
mypy-extensions==1.0.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
tzdata==2024.2
